+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- Removed buffer pooling from GRPC outbound requests which had possible data
+  corruption issues.
 
 ## [1.27.1] - 2017-01-22
 ### Changed